Always publish log messages to both .../logs and .../logs/<level> topics
... regardless of the configured MQTT protocol version (v5 or v3.x).

Also include an MQTT v5 user-property ("time") with the message's origination
time/date stamp in ISO-8601 format in all log messages (to both topics).

This will at least double the amount of MQTT log traffic to the broker. When
MQTT v5 is used, the additional user-properties will increase it even a little
more.

Is this what you were suggesting Ree?