Fix "error 13 opening" files when using init-fidonet.js
I finally reproduced this issue myself on a fresh install on Windows.

We needed explicit file closing here as going out of scope doesn't immediately
finalize (and close) a File object. That may not happen until garbage
collection and thus prevent subsequent re-opens of the same files, so always
close() your files.

init-fidonet.js calls install-binkit.js which uses cfglib.js, so that was
a source of these errors too.