Add support for ctrl/text.ini, used for sysop-override of text.dat strings
The syntax of text.ini is:
<id>: "string"
<id>: "string"

Where <id> is the named identifier of a text.dat string as seen ctrl/text.dat
and string is a literal string value (optionally, in double-quotes). Line
continuations (for long strings) are not supported: you'll just need to put
your string all on one long line (up to 1024 char string).
C-escape characters for control chars and hex-encoded chars are supported.

This should address feature request issue #644