Extend max language code (user 'lang' property) length from 3 to 8 chars
Ragnarok (DOCKSUD) would like to use lang_COUNTRY for their language files
as  "I use es_AR and is exactly not same to es_ES"

It's trivial to extend the length more if there's a desire to do that.
That's a nice thing about the new (in v3.20) user.tab file format.