Fixes for narrow terminals
In terminals < 80 cols in width, only the message # and "from" field are shown
in listings, so don't show column headings for fields that aren't shown.
Display a '-' in the flag column for deleted messages (that aren't flagged).
Shorten the prompt when listing msgs for terminals narrower than 64 cols
(for phigz). I don't know of any systems/terminals that use 64 cols however.