Fix bug that caused a [msglist:sub] last_read_mail key to be created
This just generated unnecessary noise in data/user/*.ini files, no harm.

iniSetObject (used by userprops.js) will set undefined properties as blank
keys in the section (oops, I don't recall if this is intentional or not), so
we need to 'delete' the property instead of assigning undefined to it.