Don't modify the compliance level of a cert when adding it.
If this solves the issue, it means that all certs for a specific
scfg will share the same compliance level.  For the mail server,
this level is OBLIVIOUS, so if it smells like a cert, it's considered
valid.

If this doesn't solve the issue, it needs to be backed out.