Start hacking up an sftp server.
Just a proof-of-concept for right now, there's a couple major changes
and some minor changes that will need to be made before this can
actually be enabled.

Not really testable without some seriously terrible hacks right now,
but it has been verified with the sftp client that ships with openssh
(which is based on a 20-year-old internet draft that is both expired
and superceded).