Start of integrating sftp with SyncTERM
Currently disabled, but library is built and linked.
Initial usage will be to write public key to remote... interestingly,
I think Synchronet may currently disconnect users if they attempt to
create an sftp channel, so that'll require some kind of magic to ensure
SyncTERM still works on current version of Synchronet, but still
try to auto-update on future ones.