Don't clear the console-abort flag upon return from executing a JS module
This completes the revert of commit c22063f9 from (Jun-2-2016). The Baja
part of that commit was reverted in commit 932fae30 (Nov-15-2016). This
behavior has proven to be surprising and annoying for JS mod developers
(e.g. the yesnobar.js and Nightfox's file searcher/scanner).

If we still need clearing of abort status after running JS "externals",
that should probably be done in exec_xtrn(), not here.