Restore ftpalias.cfg support for MLSx commands
Support was accidentally removed as part of commit 22e8d2e3, meaning
the ftpalias.cfg wasn't used at all in responses to the MLSx commands as
reported by Max (WESTLINE) using Total Commander, FileZilla and Directory
opus. The traditional "LIST" commands still worked fine with ftpalias.cfg
contents just fine.

The reason this code was accidentally removed was due to the errant
copy/pasted comparison with startup->html_index_file that did not belong here
and made the block appear related to HTML index file generation. It was not.
This was just a bug in the initial implementation of MLSx support in
commit d4deb4b3.

Also included in this commit:
- Return the date/size of the user's QWK packet file, if it exists, in MLSx
  response.
- send_mlsx_entry() won't report negative time_t values as file modify dates
 (flength returns -1 upon failure/file-not-found).
- get_owner_name() returns the string, making it easier to use in function
  calls.

There appears to still be some work to do to make the MLSx commands fully
compliant with RFC 3659 (e.g. "mlst filename" from ftpalias.cfg should work
but does not), but at least FileZilla displays ftpalias.cfg files and
directories correctly now.