Disable TLS 1.3
It turns out that Cryptlib appears to not support the
Hello Retry Request which means clients can't attempt to connect
unless they already know what the server supports (especially since
cryptlib doesn't support the suggested X22519 and X448 groups).

This also means it doesn't support clients that send an empty vector
in the key_share extension, which is explicitly allowed by RFC8446.

This means Cryptlib does not actually support TLS1.3, but only a
subset, and it's incapable of supporting clients that don't already
know the details of the server... which is surprising since the README
and comments in the code complain fairly bitterly about this very issue.

This was discovered since the ZNC bouncer pre-shares only an X25519
key exchange, and Cryptlib simply fails to negotiate a connection.

While actually sending the HelloRetryRequest would be a trivial patch,
maintaining the handshake hash looks non-trivial, so we'll just limp
along with the 15-year-old protocol with known vulnerabilities.