Fix rwlock_unlock() and rwlock_rdlock() recursion.
It would be nice to use TLS for the recusrion counter, but on Windows,
it looks like TLS in threads created after the object is created is
uninitialized, so we would need some kind of thread start hook thingie
which just seems like too much work.

Instead, rely on the old standby single-linked list to track.