Save disk space at the expense of CPU usage.
Previously, we compiled in a RGB -> YCbCr table and used that for
scaling, which added 128MB to the size of binaries that enabled
non-integer scaling.  The decision was then made to eliminated
non-integer scaling support from the Synchronet utilities to save
disk space.

With the use of the YCoCg-R, we can more efficiently transform between
the colour spaces, (about 50% more CPU than using the table), so
it's no longer prohibitive to do this in real-time.

This should also have the effect of making the Synchronet utilities
use the correct aspect ratio for the various screen modes rather
than forcing the use of square pixels.  I expect DigitalMan will
hate that.