Resolve full path to specified ctrl directory, before chdir
If a relative path to the ctrl directory is specified on the command-line
or in the SBBSCTRL environment variable, the loading of ctrl/*.ini would
fail after the changing of the working directory.

We should probably get replace all use of FULLPATH with _fullpath someday,
but for now, I just went with majority-rule.