Separate "natural" (no sort) order from "Date ascending" sort order for dirs
Since it is possible for someone to change the add/import date of a file
after it's been added to the filebase, the natural/unsorted order might not
actually be the same as "date ascending", so let the sysop/script-author
really re-sort by date ascending if that's what they're trying to do.

I think this is addresses the issue that Nightfox and Nelgin were describing
in #synchronet.