As it turns out, setting high DPI awareness in the process fails.
Presumably, "something" is setting it earlier so it can't be set
again because reasons.  To work around this, add a manifest to
ciolib library that asserts the binary is DPI aware and requests
Windows not do GDI scaling.

Not sure if manifest .o files in a lib are linked in by things, but
SyncTERM pulls the .o out and explicitly links it, so this should
fix SyncTERM at least.