Display warning dialog when changing cmdshell or xtrn editor's internal codes
Changing a command shell or an external editor's internal code will revert
the preference of any user's that have selected that shell or editor as their
default/preference. A sysop may not realize this, so warn them with lots of
detail in the help text if they need it. See issue #721 for details.

Also updated the help text for command shell internal codes since it still
only mentioned Baja/.bin modules.