Allow UTF-8 sequences in message subjects (passing K_UTF8 to getstr)
This was the behavior for all string input in SBBS until recently
(Commit 5dec1519a290b08b), with this change, an exception is being
made for message subjects, but the vast majority of other string
input is still limited to single-byte (i.e. US-ASCII or CP437)
characters.

This should fix issue #720