Catch when cryptFlushData() says the channel is closed.
Also, fix an unlikely race condition.  This, combined with the
cryptlib patches, fixes various weird SSH hangs resulting from
partial reads of the SSH packet headers... a rare event that's
very hard to trigger.  Special thanks to MeaTLoTioN for reporting
this, and running a BBS that manages to trigger it reliably.

Check out The Quantum Wormhole and say thanks!