Remove the mingw32 builds
mingw32 has been dead for years, and is missing huge swaths of the
Windows APIs.  strerror_s() was the straw that broke the camels
back here, but there's a lot of hackery for it.

I'm not going to go through and actually remove the build support
yet since I may want to do the 1.2 SyncTERM release with it (though
likely not), but it's not long for this world.