Don't send NodeLoggedOn message to any nodes if user isn't being logged-on
If user is being kicked off for being on "More than one node at a time", then
don't send a message to other any other nodes that the user logged-on
successfully (they haven't). Previosuly, it'd send messages to all the
lower-numbered nodes that were in-use (but not by the same user).

Reported by Nelgin.