Add support for /vfiles in sftp
The /vfiles root uses the configured virtual paths as in the FTP
and web servers.  These paths should be shorter than the ones used
in /files, and also have the benefit of translating slashes so
users will not need to type a Big Solidus if they're trying to
type long lib or dir names.