Fix for SourceForge issue 134
Translation from mouse coordinates to screen coordinates had issues
with internal vs. external scaling and for when there are gutters
(ie: black bars) on the edges.

This change snaps all gutter clicks to the nearest position inside
the window instead of discarding them, and ensures the calculations
are correct regardless of scaling mode.

Win32GDI mode was not tested, the only change there was to stop
ignoring mouse positions in the gutters.