Default SFTP Public Key to false.
It causes connections to Mystic BBSs to fail, and assumes the
server is running OpenSSH (or Synchronet, which is purposefully
compatible with OpenSSH).

I love the feature, but it's not good enough for on-by-default.

After 1.2 is released, and I do a proper sftp client, I'll likely
remove the option anyway, and just have a way to upload it on
demand.