Default to building for the current macOS version, fix 3rdp
Now if you don't specify MIN_MAC_OSX_VERSION, it will use the
"current" version of macOS as the target.  "Current" is defined as
X.0 where the major version (X) is greater than 10 and X.Y when
major version is 10.

This allows building for a specific target, but uses the newest
libraries if you don't specify.