The problem with Bullseye was the 64-bit kernel.
The configure script was using uname -m as the build platform, but
Bullseye on RPi uses a 64-bit kernel (aarch64 uname -m) with a 32-bit
userland (arm gcc build host), which resulted in a lot of bad decsions
around JIT stuff.

With this change, we explicitly pass the build/host/target values as
taken from the compilers in question, which fixes the issue for Bullseye
(But may break mingw32?  Well we'll see how the pipes go).