Add 2 new .ini settings: MainLoopDelay and SocketSelectTimeout (default: 0)
- MainLoopDelay - in milliseconds
  Set this to a non-zero value to add some CPU yielding to the main loop
- SocketSelectTimeout - in milliseconds
  Set this to a non-zero value to add some CPU yielding while waiting for
  receive data from the TCP socket

These are 2 knobs created for Fzf (FQBBS) to "to reduce the CPU usage". Their
default values are currently 0, so you'll need to set these to non-zero values
to play with each of them and see their inpact on CPU utilization and
responsiveness. I'd suggest starting with low values (e.g. 1) and experimeting
from there.

Incremented version to 0.5