Default FTN charset to CP437 (not ASCII) for QWK-imported messages
A QWKnet tagline usually contains CP437 character 254 and if/when appended to
a message during export, would "convert" the message to CP437 thus making the
"ASCII" charset advertisement incorrect.

If a FIDO CHARSET value was already specified (e.g. in HEADERS.DAT), that
value will (still) not be overridden. So the *original* import via QWK (before
it gets exported to a QWKnet) is where this change will take effect.

Fix for issue #741