Introduce system.git_time
This is the time_t (seconds since Unix epoch, Jan-1-1970 UTC) date/time of the
current (most recent) Git commit used to build the running binary.

If a JS developer wants to check if a build is recent-enough to include some
change, this is the property they should use to check (with >= comparison).