When server_binary option is set, request remote BINARY_TX option
As reported by Fzf (FQBBS):

It addresses the local configuration but unfortunately it still doesn't set
remote options.  The remote is usually going to be in binary mode but SVDM has
the remote option set to ASCII by default.  A CR from the remote then gets held
up until a second byte is sent.

Also suggested by Fzf, when server_binary mode is not configured, then advertise
and request binary mode is/be disabled, explicitly.