Xsword's Oblivion/2 Clone Command Shell from 1994
Although the FILE_ID.DIZ suggests this is a clone of Oblvivion/2 v2.55,
the latest Oblivion/2 appears to be v2.40.

This shell source (in Baja) is a 99% copy of Yojimbo's (The Dojo BBS) Vision-2
shell source code from earlier in 1994.

The menu/display files need a little work, so they'll be committed separately.

I mainly wanted to commit this and then record the history of modifications
necessary to "modernize" the shell and at least get it compiling and useful
with recent versions of Synchronet BBS for Win32 and *nix and as demonstrated
at https://youtu.be/HlwQ0uX4S04