Re-assign the dirnum and subnum fields after sorting dirs or subs
The dirnum and subnum are used for the next/previous (left and right) arrow
nav through sub-boards and directories. For message groups or file libs that
are configured to auto-sort the sub-boards or directory list, when importing
an unsorted list, the dirnum and subnum values would then be wrong after the
qsort() at the end of the import thus making the left/right nav thing broken
(jump to other groups/libs and such).

This fixes that.

Could I have fixed this by passing an index value to the next/prev_dirnum()
and _subnum() functions and returning the next/previous index value instead of
relying on the subnum/dirnum elemment to be correct? Maybe. <shrug>
Do I like second guessing myself? Not really.