Use name as default address when creating new TCP dialing entry
I usually use the hostname for the name of the dialing entry and got tired
of always having to type the hostname twice when creating a new entry.

If the name looks like a hostname or IP address (includes a dot but
no spaces) and the connection type is appropriate, make the name the
default string for the address (easily over-ridden if wrong, by just
typing the correct address). This time-saving feature doesn't work for
single-word hostnames (e.g. "localhost") or IPv6 addresses.