Use the telnet gateway proxy (not pass-through) by default
Pass-through only really does anything useful if the BBS user's client is a
Telnet client and it responds as the server expects. Since the user's
Telnet client (when using one) has already negotiated Telnet options, they
won't ACK options that the server expects them to, so this doesn't really
work. The proxy fixes that issue.