Log the current connection protocol as a separate property in logon.jsonl
Accession reported that (sometimes?) his NNTP logins show up in his
logon.jsonl files. This is most likely just a race condition when logging
in close to the same time with multiple protocols (e.g. Telnet and NNTP
in this case). So, let's log the client.protocol as a separate JSON property
and not require on the user.connection property (which is overwritten for
*every* login, for almost all connection/server types, not just the terminal
server). The client.protocol is always correct for the current session running
the script.

An NNTP login should never appear in the logon.json file.