Open the quote file in "binary" mode (only applicable on Windows)
If the quote file contained a Ctrl-Z (ASCII 26) char, it would be truncated
at that point. Some UTF-8 messages that include a "right arrow" unicode code
point are translated to ASCII Ctrl-Z (ASCII 26) char, which is interpretted
by Windows as "EOF" (end of file) marker for files open in "text mode" and
the file won't be read beyond that char. Open the file in binary mode to
more closely match the *nix fopen() before.