Add suport for option/item tagging using the WIN_TAG ulist() mode flag
For use with fileman.js:

The space bar can be used by the user to "tag" items, or Ctrl-A to toggle
tag status for *all* items. Returns MSK_TAG|with the option tagged/untagged
or MSG_TAGALL when all tags are being toggled (via ^A).

Also, fix bug where ^Find command would not restore the status bar after
ulist() would overwrite it with the string-input/edit hot-keys available.