FileBase.add() now supports override to *never* extract/use DIZ if desired
Working on fileman.js, when moving files to another base, the output of unzip
(trying to extra a FILE_ID.DIZ) was corrupting the display. We don't really
want a DIZ file to get imported when moving files between bases anyway.

So the second parameter to FileBase.add() can be either:
- undefined: will try to extract/use DIZ only if there's no ext desc already
- true: will try to extract/use a DIZ even if there's an ext desc already
- false: will never try to extract/use a DIZ

If the directory is not configured for DIZ support, then the DIZ extraction/use
will never occur under any condition.