Fix 64-char filename length limit for FileBase.get[_path/size/time] methods
This was only a problem when these methods were passed a file-meta-object
instead of the file name. In hindsight, that's kind of a dumb API "feature"
since we're just parsing the name property out of the object.

The file_t.file_idx.name is limited to 64 chars, so don't use that struct
element to store the parsed "name" property of the passed file-meta-object.