Introduce a restore() uifc_api method for restoring a saved window buffer
This allows for menu options on an "active" list (with both WIN_SAVE and
WIN_ACT mode flags) to close the window and restore the most recently saved
window.

Also, for the range check on the save buffer count in ulist(), if we've
hit the bounds, don't *just* beep (though, that was useful), but turn off the
WIN_SAV mode flag so we don't overflow the sav[] array and corrupt the heap!