Use the new uifc_t.restore() method when decrementing the save_num
This is the proper way to manually-restore the most recently saved
window (via the WIN_SAV mode flag to the list() method).

I wanted something that was backward comaptible with earlier builds
of SBBS, so just doing a magic thing when decrementing the save_num
rather than adding a new method seemed the seamless way to do it.