Add new "Search Pending Files" operation
Sort of the opposite of the "Offline Files" search, find files copied into
the directory storage paths but not yet added to the filebase, with a
"Add File..." option (and tagging support) along with DELete files option.

Auto-ignore listing files (e.g. FILES.BBS, etc.) when searchign for new
files to add to a directory's filebase.

If an argument is passed to fileman.js, use that as the UIFC mode (for testing
purposes only).

Fix case-sensitive file compares (on *nix OSes) when adding files (now
case-insensitive).

Fix issue with get file paths for files not yet in a filebase (using
FileBath.get_path unnecessarily).

Display virtual paths to files in many place, when available (> SBBS v3.19).